cleanup of shape evaluation code
