+++ Examples/Python/ellipse_evaluate.py
