+++ Libs/Particles/ShapeEvaluation.h
@@ -7,13 +7,15 @@
 #include "ParticleSystem.h"
 
 namespace shapeworks {
+
+template <unsigned int VDimension=3>
 class ShapeEvaluation {
 public:
     ShapeEvaluation();
 
     static double ComputeCompactness(const ParticleSystem &particleSystem, const int nModes, const std::string &saveScreePlotTo="");
     static double ComputeGeneralization(const ParticleSystem &particleSystem, const int nModes, const std::string &saveTo="");
-    static double ComputeSpecificity(const ParticleSystem &particleSystem, const int nModes);
+    static double ComputeSpecificity(const ParticleSystem &particleSystem, const int nModes, const std::string &saveTo="");
 };
 }
